chlorobot/requirements.txt
asgiref==1.1.2
attrs==17.2.0
autobahn==17.10.1
Automat==0.6.0
certifi==2017.11.5
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
daphne==1.3.0
Django==1.11.7
django-channels==0.7.0
django-db-logger==0.1.6
django-filter==1.1.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
oauthlib==2.0.6
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
zope.interface==4.4.3
